requirements.txt
fastapi
uvicorn[standard]
asyncpg
python-dotenv
email-validator
